fix logging format

+++ src/dftracer/utils/common/logging.h
@@ -8,6 +8,8 @@
 #include <chrono>
 #include <cstdio>
 #include <string>
+#include <cstdint>
+#include <cinttypes>
 
 namespace dftracer::utils {
 inline std::string dftracer_utils_macro_get_time() {
@@ -20,9 +22,9 @@ inline std::string dftracer_utils_macro_get_time() {
     auto now = std::localtime(&dftracer_utils_ts_t);
     char dftracer_utils_ts_time_str[256];
     snprintf(dftracer_utils_ts_time_str, sizeof(dftracer_utils_ts_time_str),
-             "%04d-%02d-%02d %02d:%02d:%02d.%lld", now->tm_year + 1900,
+             "%04d-%02d-%02d %02d:%02d:%02d.%03" PRId64, now->tm_year + 1900,
              now->tm_mon + 1, now->tm_mday, now->tm_hour, now->tm_min,
-             now->tm_sec, dftracer_utils_ts_millis);
+             now->tm_sec, static_cast<std::int64_t>(dftracer_utils_ts_millis));
     return dftracer_utils_ts_time_str;
 }
 }  // namespace dftracer::utils
